requirements.txt
astroid==2.3.3
Click==7.0
coverage==5.0
Flask==1.1.1
Flask-Cors==3.0.8
ipaddress==1.0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
lxml==4.4.2
mail-parser==3.9.3
MarkupSafe==1.1.1
mccabe==0.6.1
nose2==0.9.1
pip-autoremove==0.9.1
pylint==2.4.4
python-docx==0.8.10
rope==0.14.0
simplejson==3.17.0
six==1.12.0
Werkzeug==0.16.0
wrapt==1.11.2
